Improved quality of error checking
An exception should now be thrown if too few or too many matches for a
particular team are entered in games.txt, hopefully making typos in
match data easier to find. This was supposed to happen in previous
versions but did not.

Team.java now contains getOpponents, that returns the number of
opponents registered to the team. updateSOS no longer attempts to (and
fails) to check all teams have the same number of opponents.

StrengthOfSchedule.java now includes a check to ensure each team has the
same number registered so errors in match history input are easier to
spot.